// src/app/pages/quizComp/models/questions.ts
import {Choice } from './choice';

export class Questions {
    question: string;
    id: number;
    choices: string;
    answer: string;
    ability: number;
    difficulty: number;

    // constructor(data:any){
    //     this.choices = [];
    //     data.options.forEach(o => {
    //         this.choices.push(new Choice());
    //     });
    // }
}